+++ src/engine/cameras/camera_manipulation.js
