+++ src/engine/cameras/camera_shake.js
