+++ src/engine/core/loop.js
