+++ src/engine/core/resource_map.js
