+++ src/engine/core/shader_resources.js
