+++ src/engine/game_objects/game_object.js
