+++ src/engine/game_objects/game_object_set.js
