+++ src/engine/renderables/font_renderable.js
