+++ src/engine/renderables/line_renderable.js
