+++ src/engine/renderables/sprite_animate_renderable.js
