+++ src/engine/renderables/sprite_renderable.js
