+++ src/engine/renderables/texture_renderable.js
