+++ src/engine/renderables/texture_renderable_main.js
