+++ src/engine/resources/default_resources.js
