+++ src/engine/resources/text.js
