+++ src/engine/resources/xml.js
